+++ backend/src/main/java/com/hades/foxtube/dao/CommentDao.java
@@ -1,9 +1,8 @@
 package com.hades.foxtube.dao;
 
 import com.hades.foxtube.model.Comment;
-import org.apache.ibatis.annotations.Mapper;
-
 import java.util.List;
+import org.apache.ibatis.annotations.Mapper;
 
 /**
  * @Author: Hades @Date: 2024/5/29 @Description:
@@ -12,7 +11,7 @@ import java.util.List;
 public interface CommentDao {
   void insertComment(Comment comment);
 
-  List<Comment> getComments(Integer videoId, Integer offset, Integer limit);
+  List<Comment> getComments(Long videoId, Long offset, Long limit);
 
-  Integer getCommentCount(Integer videoId);
+  Long getCommentCount(Long videoId);
 }
