+++ backend/src/main/java/com/hades/foxtube/dao/VideoDao.java
@@ -1,20 +1,19 @@
 package com.hades.foxtube.dao;
 
 import com.hades.foxtube.model.Video;
-import org.apache.ibatis.annotations.Mapper;
-
 import java.util.List;
+import org.apache.ibatis.annotations.Mapper;
 
 /**
  * @Author: Hades @Date: 2024/5/29 @Description:
  */
 @Mapper
 public interface VideoDao {
-  List<Video> getVideos(Integer authorId, Integer offset, Integer limit);
+  List<Video> getVideos(Long authorId, Long offset, Long limit);
 
   void insertVideo(Video video);
 
-  Integer getVideoCount(Integer authorId);
+  Long getVideoCount(Long authorId);
 
-  Video getVideo(Integer id);
+  Video getVideo(Long id);
 }
