+++ backend/src/main/java/com/hades/foxtube/service/VideoService.java
@@ -1,18 +1,17 @@
 package com.hades.foxtube.service;
 
 import com.hades.foxtube.model.Video;
-
 import java.util.List;
 
 /**
  * @Author: Hades @Date: 2024/5/29 @Description:
  */
 public interface VideoService {
-  List<Video> getVideos(Integer authorId, Integer offset);
+  List<Video> getVideos(Long authorId, Long offset);
 
   void insertVideo(Video video);
 
-  Integer getVideoCount(Integer authorId);
+  Long getVideoCount(Long authorId);
 
-  Video getVideo(Integer id);
+  Video getVideo(Long id);
 }
