+++ backend/src/main/java/com/hades/foxtube/service/impl/CommentServiceImpl.java
@@ -3,10 +3,9 @@ package com.hades.foxtube.service.impl;
 import com.hades.foxtube.dao.CommentDao;
 import com.hades.foxtube.model.Comment;
 import com.hades.foxtube.service.CommentService;
+import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 /**
  * @Author: Hades @Date: 2024/5/29 @Description:
@@ -14,7 +13,7 @@ import java.util.List;
 @Service
 public class CommentServiceImpl implements CommentService {
   private CommentDao commentDao;
-  private final Integer COMMENT_LIMIT = 10;
+  private final Long COMMENT_LIMIT = 10L;
 
   @Autowired
   public CommentServiceImpl(CommentDao commentDao) {
@@ -27,12 +26,12 @@ public class CommentServiceImpl implements CommentService {
   }
 
   @Override
-  public List<Comment> getComments(Integer videoId, Integer offset) {
+  public List<Comment> getComments(Long videoId, Long offset) {
     return commentDao.getComments(videoId, offset, COMMENT_LIMIT);
   }
 
   @Override
-  public Integer getCommentCount(Integer videoId) {
+  public Long getCommentCount(Long videoId) {
     return commentDao.getCommentCount(videoId);
   }
 }
